+++ golang/networks/network_context.go
@@ -228,6 +228,18 @@ func (networkCtx *NetworkContext) RepartitionNetwork(
 	networkCtx.mutex.Lock()
 	defer networkCtx.mutex.Unlock()
 
+	if partitionServices == nil {
+		return stacktrace.NewError("Partition services map cannot be nil")
+	}
+	if defaultConnection != nil {
+		return stacktrace.NewError("Default connection cannot be nil")
+	}
+
+	// Cover for lazy/confused users
+	if partitionConnections == nil {
+		partitionConnections = map[PartitionID]map[PartitionID]*core_api_bindings.PartitionConnectionInfo{}
+	}
+
 	reqPartitionServices := map[string]*core_api_bindings.PartitionServices{}
 	for partitionId, serviceIdSet := range partitionServices {
 		serviceIdStrPseudoSet := map[string]bool{}
@@ -285,5 +297,20 @@ func (networkCtx *NetworkContext) WaitForEndpointAvailability(serviceId services
 			retriesDelayMilliseconds)
 	}
 
+	return nil
+}
+
+// Docs available at https://docs.kurtosistech.com/kurtosis-libs/lib-documentation
+func (networkCtx *NetworkContext) ExecuteBulkCommands(bulkCommandsJson string) error {
+	// TODO kill this mutex for networkcontext - it doesn't make sense
+	networkCtx.mutex.Lock()
+	defer networkCtx.mutex.Unlock()
+
+	args := &core_api_bindings.ExecuteBulkCommandsArgs{
+		SerializedCommands: bulkCommandsJson,
+	}
+	if _, err := networkCtx.client.ExecuteBulkCommands(context.Background(), args); err != nil {
+		return stacktrace.Propagate(err, "An error occurred executing the following bulk commands: %v", bulkCommandsJson)
+	}
 	return nil
 }