+++ scripts/regenerate-protobuf-output.sh
@@ -23,7 +23,7 @@ if ! GO_MOD_FILEPATH="${root_dirpath}/${GOLANG_DIRNAME}/go.mod" "${GENERATOR_SCR
 fi
 echo "Successfully generated Go bindings in directory '${go_output_dirpath}'"
 
-#TypeScript
+# TypeScript
 typescript_output_dirpath="${root_dirpath}/${TYPESCRIPT_DIRNAME}/core_api_bindings"
 if ! "${GENERATOR_SCRIPT_FILENAME}" "${input_dirpath}" "${typescript_output_dirpath}" typescript; then
     echo "Error: An error occurred generating TypeScript bindings in directory '${typescript_output_dirpath}'" >&2